+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Removed the `raceN` and `awaitN` functions from the `promise` module.
 - Removed the `map` module.
+- Removed the `type_of` function and `TypeOf` type from the `javascript` module.
 
 ## v0.13.0 - 2024-08-30